refactor: edit directory route, add component Star
