+++ client/src/components/Main/Introduction.tsx
